pyproject.toml
[project]
name = "nldf-factory-simulation"
version = "0.1.0"
description = "NLDF Factory Simulation with Agent Evaluation"
requires-python = ">=3.11"
dependencies = [
    "simpy",
    "paho-mqtt",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "httpx",
    "openai-agents>=0.2.3",
    "openai>=1.12.0",
]

[project.scripts]
eval = "eval:main"
run-simulation = "run_multi_line_simulation:run_simulation_multi"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8",
]